feat(图表): 支持矩形树图
